WIP Naive embedding server to MSM
Tried to embed server to state machine.
Injecting server to MSM::MachineStateManager looks like
nonsense because it is not responsible for state transitions
in EM, so I tried passing it as parameter to States, but it
requires state factory signature to change. So, there should be
a better way to pass server in MSM.